Reorganize forms and validators
