+++ app/services/idv/steps/ipp/address_step.rb
@@ -5,7 +5,7 @@ module Idv
         STEP_INDICATOR_STEP = :verify_info
 
         def call
-          Idv::InPersonProofingAddressForm::ATTRIBUTES.each do |attr|
+          Idv::InPerson::AddressForm::ATTRIBUTES.each do |attr|
             flow_session[:pii_from_user][attr] = flow_params[attr]
           end
         end
@@ -19,8 +19,8 @@ module Idv
         private
 
         def form_submit
-          Idv::InPersonProofingAddressForm.new.submit(
-            permit(*Idv::InPersonProofingAddressForm::ATTRIBUTES),
+          Idv::InPerson::AddressForm.new.submit(
+            permit(*Idv::InPerson::AddressForm::ATTRIBUTES),
           )
         end
       end
