--- app/validators/idv/form_in_person_proofing_address_validator.rb
@@ -1,11 +1,0 @@
-module Idv
-  module FormInPersonProofingAddressValidator
-    extend ActiveSupport::Concern
-    include FormAddressValidator
-
-    included do
-      validates :same_address_as_id,
-                presence: true
-    end
-  end
-end
