+++ app/validators/idv/in_person/form_address_validator.rb
@@ -1,0 +1,13 @@
+module Idv
+  module InPerson
+    module FormAddressValidator
+      extend ActiveSupport::Concern
+      include Idv::FormAddressValidator
+
+      included do
+        validates :same_address_as_id,
+                  presence: true
+      end
+    end
+  end
+end
