+++ app/views/idv/in_person/address.html.erb
@@ -77,7 +77,7 @@
             as: :radio_buttons,
             wrapper: :borderless_uswds_radio_buttons,
             label: t('in_person_proofing.form.address.same_address'),
-            required: true,
+            # required: true,
             collection: [
               [t('in_person_proofing.form.address.same_address_choice_yes'), true],
               [t('in_person_proofing.form.address.same_address_choice_no'), false],
